IBM has contributed an howto for creating PDO drivers.
They supplied the text to me as a word doc, and I've converted that to
Open Office and then used OOo2sDbk (http://www.chez.com/ebellot/ooo2sdbk/)
to arrive at what you see here.

This text is based on an older version of the driver API, and I will be
updating this in the near future.

I'd like to ask a phpdoc'er to volunteer to patch up the formatting to make
it acceptable for inclusion in the manual before I do this; I'm really short
on time right now, and I'm not up to speed with the latest phpdoc rules.

Feel free to split this document out into multiple files as you see fit.

Many thanks!


git-svn-id: http://svn.php.net/repository/phpdoc/en@196270 c90b9560-bf6c-de11-be94-00142212c4b1