Change the documentation of sqlite_key() to reflect that it can only be called as an instance method, unlike every other SQLite function. The previous changelog was, in fact, completely incorrect.
This is absolutely not the right way to do this, and anyone who has a better
idea for how to represent this case is welcome to commit their own fix.
Unfortunately, because the OO API is just a set of method aliases for the
procedural API and it's infrequently used, we don't have the class/method
structure that we have in (say) mysqli, and attempting to retrofit it now would
result in all of the URLs for the SQLite function pages breaking.

Ugh.

Fixes doc bug #54073 (The sqlite_key() documentation should mentioned
procedural usage), for a very low value of fixes.


git-svn-id: http://svn.php.net/repository/phpdoc/en@328319 c90b9560-bf6c-de11-be94-00142212c4b1