Clarify the uniqueness and stability of casting resources to string/integer.
It was pointed out on comp.lang.php that the existing wording was unclear on
the point of what in the resource->string casting behaviour shouldn't be relied
upon (whether it was not guaranteed that the string is unique, or just that the
exact format of the string is subject to change). I've also added a paragraph
to the integer page noting what a resource->integer cast results in.


git-svn-id: http://svn.php.net/repository/phpdoc/en@332894 c90b9560-bf6c-de11-be94-00142212c4b1