Update version information to include PHP 7.
This was mostly done programmatically, and is likely wrong for a small number
of functions. We may have to wait for bug reports, unless someone wants to
review this by hand. Basic efforts were made to avoid adding "PHP 7" to
functions that were removed in PHP 5.x (x > 0) and PHP 7, but there are
undoubtedly still issues.

In case we're still using Docbook for PHP 8, I've uploaded the script that I
wrote to do this to GitHub at
https://github.com/LawnGnome/php-doc-version-migrator.


git-svn-id: http://svn.php.net/repository/phpdoc/en@338406 c90b9560-bf6c-de11-be94-00142212c4b1