Begin beefing up our warnings for things removed in 7.0.
First up: ereg.

I've added a couple of new snippets based on the MySQL snippets that allow us
to provide alternatives in the big red warning box atop each page, since that
approach seems to work well there. I'll add more variations as required as I
work through the various extensions that need to be updated.

In general, I'd like to use this approach more consistently moving forward as
we deprecate and remove things, since it gives the user a very obvious link to
follow to find out about the new way of doing things.


git-svn-id: http://svn.php.net/repository/phpdoc/en@338425 c90b9560-bf6c-de11-be94-00142212c4b1