JSON_BIGINT_AS_STRING is NOT enabled by default.
https://3v4l.org/4IZl6
https://github.com/php/php-src/blob/master/ext/json/tests/008.phpt

Maybe there was a confusion: the doc originally read "(...) supported options. The first is JSON_BIGINT_AS_STRING that allows casting big integers to string instead of floats which is the default. (...)", but here "which is the default" was referring to "[casting big integers to] floats", not to the option.

-- 
Provided by anonymous 94624 ()

git-svn-id: http://svn.php.net/repository/phpdoc/en@345972 c90b9560-bf6c-de11-be94-00142212c4b1